iOS 14.0

+++ iOS/Frameworks/CFNetwork.framework/__NSCFURLLocalStreamTaskWorkRead.h
@@ -8,8 +8,8 @@
 
 @interface __NSCFURLLocalStreamTaskWorkRead
 {
-    unsigned long long _minBytes;
-    unsigned long long _maxBytes;
+    unsigned int _minBytes;
+    unsigned int _maxBytes;
     CDUnknownBlockType _completion;
     NSData *_readData;
     _Bool _eof;
