+++ iOS/PrivateFrameworks/GameCenterFoundation.framework/NSDictionary-GKCollectionUtils.h
@@ -15,7 +15,7 @@
 - (id)_gkSubDictionaryWithKeys:(id)arg1;
 - (id)_gkDictionaryByRemovingObjectForKey:(id)arg1;
 - (id)_gkDictionaryByRemovingObjectsForKeys:(id)arg1;
-- (id)_gkDescriptionWithChildren:(int)arg1;
+- (id)_gkDescriptionWithChildren:(long long)arg1;
 - (id)_gkImageURLForSize:(long long)arg1 scale:(double)arg2;
 - (id)_gkImageURLForSize:(long long)arg1 scale:(double)arg2 foundSize:(out unsigned long long *)arg3;
 @end
