+++ iOS/PrivateFrameworks/GameCenterFoundation.framework/NSSet-GKCollectionUtils.h
@@ -13,6 +13,6 @@
 - (id)_gkMapDictionaryWithKeyPath:(id)arg1 valueKeyPath:(id)arg2;
 - (id)_gkMapDictionaryWithKeyPath:(id)arg1;
 - (id)_gkMapWithBlock:(CDUnknownBlockType)arg1;
-- (id)_gkDescriptionWithChildren:(int)arg1;
+- (id)_gkDescriptionWithChildren:(long long)arg1;
 @end
