+++ iOS/PrivateFrameworks/MapsSupport.framework/SYSessionDelegate-Protocol.h
@@ -12,7 +12,7 @@
 - (_Bool)syncSession:(SYSession *)arg1 resetDataStoreWithError:(id *)arg2;
 - (void)syncSession:(SYSession *)arg1 didEndWithError:(NSError *)arg2;
 - (void)syncSession:(SYSession *)arg1 applyChanges:(NSArray *)arg2 completion:(void (^)(_Bool, NSError *))arg3;
-- (long long)syncSession:(SYSession *)arg1 enqueueChanges:(_Bool (^)(id <SYChange>))arg2 error:(id *)arg3;
+- (unsigned int)syncSession:(SYSession *)arg1 enqueueChanges:(_Bool (^)(id <SYChange>))arg2 error:(id *)arg3;
 
 @optional
 - (void)syncSession:(SYSession *)arg1 successfullySynced:(NSSet *)arg2;
