+++ iOS/PrivateFrameworks/TelephonyUI.framework/TPNumberPadDarkStyleButton.h
@@ -10,8 +10,8 @@
 
 + (double)highlightedCircleViewAlpha;
 + (double)unhighlightedCircleViewAlpha;
-+ (id)imageForCharacter:(unsigned int)arg1 highlighted:(_Bool)arg2;
-+ (id)imageForCharacter:(unsigned int)arg1;
++ (id)imageForCharacter:(long long)arg1 highlighted:(_Bool)arg2;
++ (id)imageForCharacter:(long long)arg1;
 - (id)buttonColor;
 - (id)defaultColor;
